Refactor payment service to remove Redis dependency
- Removed Redis integration from the payment service, including Bull queue setup and related modules.
- Updated logger service to store logs directly in the database instead of queuing.
- Adjusted application modules to reflect the removal of Bull and Redis dependencies.
- Added new Nginx configuration files for development and production environments.
