+++ application/payment-service/package.json
@@ -23,22 +23,17 @@
     "migration:revert": "typeorm-ts-node-commonjs migration:revert"
   },
   "dependencies": {
-    "@bull-board/api": "^5.10.2",
-    "@bull-board/express": "^5.10.2",
-    "@nestjs/bull": "^10.0.1",
     "@nestjs/common": "^10.0.0",
     "@nestjs/config": "^3.0.0",
     "@nestjs/core": "^10.0.0",
     "@nestjs/platform-express": "^10.0.0",
     "@nestjs/typeorm": "^10.0.0",
     "@types/express": "^4.17.17",
-    "bull": "^4.11.3",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
     "nest-winston": "^1.9.4",
     "omise": "^0.10.0",
     "pg": "^8.11.3",
-    "redis": "^4.6.7",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.8.1",
     "typeorm": "^0.3.17",
@@ -49,7 +44,6 @@
     "@nestjs/cli": "^10.0.0",
     "@nestjs/schematics": "^10.0.0",
     "@nestjs/testing": "^10.0.0",
-    "@types/bull": "^4.10.0",
     "@types/jest": "^29.5.2",
     "@types/node": "^20.3.1",
     "@types/pg": "^8.10.2",
