+++ application/payment-service/src/common/logger/logger.module.ts
@@ -1,12 +1,11 @@
 import { Module } from '@nestjs/common';
-import { BullModule } from '@nestjs/bull';
+import { TypeOrmModule } from '@nestjs/typeorm';
 import { LoggerService } from './logger.service';
+import { PaymentLog } from '../../database/entities/payment-log.entity';
 
 @Module({
   imports: [
-    BullModule.registerQueue({
-      name: 'payment-logs',
-    }),
+    TypeOrmModule.forFeature([PaymentLog]),
   ],
   providers: [LoggerService],
   exports: [LoggerService],
