+++ application/payment-service/src/payments/payments.module.ts
@@ -1,6 +1,5 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { BullModule } from '@nestjs/bull';
 import { PaymentsController } from './payments.controller';
 import { PaymentsService } from './payments.service';
 import { Payment } from '../database/entities/payment.entity';
@@ -10,9 +9,6 @@ import { LoggerModule } from '../common/logger/logger.module';
 @Module({
   imports: [
     TypeOrmModule.forFeature([Payment, Seat]),
-    BullModule.registerQueue({
-      name: 'payment-logs',
-    }),
     LoggerModule,
   ],
   controllers: [PaymentsController],
