+++ application/payment-service/src/webhook/webhook.module.ts
@@ -1,5 +1,4 @@
 import { Module } from '@nestjs/common';
-import { BullModule } from '@nestjs/bull';
 import { WebhookController } from './webhook.controller';
 import { WebhookService } from './webhook.service';
 import { PaymentsModule } from '../payments/payments.module';
@@ -7,9 +6,6 @@ import { LoggerModule } from '../common/logger/logger.module';
 
 @Module({
   imports: [
-    BullModule.registerQueue({
-      name: 'payment-logs',
-    }),
     PaymentsModule,
     LoggerModule,
   ],
