+++ nginx/nginx.dev.conf
@@ -1,0 +1,76 @@
+# nginx.dev.conf
+# This configuration is for local development environments.
+
+events {
+    # Set the number of worker connections.
+    worker_connections 1024;
+}
+
+http {
+    # Define upstream for the payment service.
+    upstream payment_service {
+        # Use round_robin for simple load balancing (default).
+        # Define the server and port.
+        server payment-service:3000;
+        # You can add more servers if you run multiple instances.
+        # server payment-service-2:3000;
+    }
+
+    # Define upstream for the frontend service.
+    upstream frontend_service {
+        server frontend:3001;
+        # server frontend-2:3001;
+    }
+
+    # Define upstream for the new backend service.
+    upstream hello_backend_service {
+        server backend-service:3000;
+    }
+
+    # Main server block for handling incoming requests.
+    server {
+        # Listen on port 80, the default HTTP port.
+        listen 80;
+
+        # Frontend application.
+        location / {
+            # Pass all requests to the frontend service.
+            proxy_pass http://frontend_service;
+            # Preserve the original Host header.
+            proxy_set_header Host $host;
+            # Preserve the original client IP address.
+            proxy_set_header X-Real-IP $remote_addr;
+            # Forward the client's IP address through the proxy.
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        }
+
+        # API routes for the payment service (all endpoints).
+        location /api/ {
+            # Pass requests to the payment service.
+            proxy_pass http://payment_service;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            # Allow larger request bodies, useful for webhook payloads.
+            client_max_body_size 10M;
+        }
+
+        # API routes for the new backend service.
+        location /hello-api/ {
+            # Pass requests to the new backend service.
+            proxy_pass http://hello_backend_service;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        }
+
+        # Health check endpoint.
+        location /health {
+            # This is a specific health check for the payment service.
+            proxy_pass http://payment_service/api/v1/health;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        }
+    }
+}
